requirements.txt
numpy==1.15.0
Pillow==7.0.0
opencv-python==4.2.0.32
tensorflow-gpu==1.15